+++ shell/powershell/winTerm.Shell/winTerm.Shell.psm1
@@ -3,7 +3,7 @@
 
 Set-StrictMode -Version Latest
 
-$script:WinTermModuleVersion = '1.2.3'
+$script:WinTermModuleVersion = '1.2.4'
 $script:WinTermProtocolVersion = 1
 $script:WinTermIntegrationEnabled = $false
 $script:WinTermPromptWrapped = $false